+++ src/main/java/ru/ravilov/ClientBase/repository/CarCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.ravilov.ClientBase.model.CarCategory;
 
 @Repository
-public interface CarCategoryRepository extends JpaRepository<CarCategory,Integer> {
+public interface CarCategoryRepository extends JpaRepository<CarCategory,Long> {
 }
